create Book component, rename Pages for clarity in editor
