+++ src/Components/Book.tsx
@@ -1,0 +1,10 @@
+import * as React from "react";
+//import styled from '../utils/theme'
+
+type BookProps = {
+  book?: {};
+};
+
+const Book: React.FunctionComponent<BookProps> = () => <div></div>;
+
+export default Book;
